simplify Windows Service implementation and revert exporter.go struct refactoring
