+++ main.go
@@ -3,15 +3,26 @@
 package main
 
 import (
+	"context"
+	"crypto/tls"
+	"crypto/x509"
 	"fmt"
+	"net"
+	"net/http"
 	"os"
-	"os/signal"
-	"syscall"
+	"sync"
+	"time"
+
+	"github.com/digineo/go-ping"
+	mon "github.com/digineo/go-ping/monitor"
 
 	"github.com/czerwonk/ping_exporter/config"
 
 	"github.com/alecthomas/kingpin/v2"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promhttp"
 	log "github.com/sirupsen/logrus"
+	inotify "gopkg.in/fsnotify.v1"
 )
 
 const version string = "1.2.1"
@@ -61,9 +72,11 @@ var (
 
 	rttMetricsScale = rttInMills // might change in future
 	rttMode         = kingpin.Flag("metrics.rttunit", "Export ping results as either seconds (default), or milliseconds (deprecated), or both (for migrations). Valid choices: [s, ms, both]").Default("s").String()
+	desiredTargets  *targets
 )
 
 func main() {
+	desiredTargets = &targets{}
 	kingpin.Parse()
 
 	if *showVersion {
@@ -118,19 +131,18 @@ func main() {
 }
 
 func runInteractive(cfg *config.Config) {
-	exporter := NewExporter(cfg)
+	globalResolver := setupGlobalResolver(cfg)
 
-	err := exporter.Start()
+	m, err := startMonitor(cfg, globalResolver)
 	if err != nil {
-		log.Fatalf("failed to start exporter: %v", err)
+		log.Errorln(err)
+		os.Exit(2)
 	}
 
-	sigChan := make(chan os.Signal, 1)
-	signal.Notify(sigChan, syscall.SIGINT, syscall.SIGTERM)
-	<-sigChan
+	collector := NewPingCollector(enableDeprecatedMetrics, rttMetricsScale, m, cfg)
+	go watchConfig(desiredTargets, globalResolver, m, collector)
 
-	log.Info("Shutting down...")
-	exporter.Stop()
+	startServer(collector)
 }
 
 func printVersion() {
@@ -140,6 +152,294 @@ func printVersion() {
 	fmt.Println("Metric exporter for go-icmp")
 }
 
+func startMonitor(cfg *config.Config, globalResolver Resolver) (*mon.Monitor, error) {
+	var bind4, bind6 string
+	if ln, err := net.Listen("tcp4", "127.0.0.1:0"); err == nil {
+		// ipv4 enabled
+		if err := ln.Close(); err != nil {
+			return nil, fmt.Errorf("failed to close tcp4 listener: %w", err)
+		}
+		bind4 = "0.0.0.0"
+	}
+	if ln, err := net.Listen("tcp6", "[::1]:0"); err == nil {
+		// ipv6 enabled
+		if err := ln.Close(); err != nil {
+			return nil, fmt.Errorf("failed to close tcp6 listener: %w", err)
+		}
+		bind6 = "::"
+	}
+	pinger, err := ping.New(bind4, bind6)
+	if err != nil {
+		return nil, fmt.Errorf("cannot start monitoring: %w", err)
+	}
+
+	if pinger.PayloadSize() != cfg.Ping.Size {
+		pinger.SetPayloadSize(cfg.Ping.Size)
+	}
+
+	if cfg.Ping.FirewallMark > 0 {
+		err := pinger.SetMark(cfg.Ping.FirewallMark)
+		if err != nil {
+			return nil, fmt.Errorf("failed to set fwmark: %w", err)
+		}
+	}
+
+	monitor := mon.New(pinger,
+		cfg.Ping.Interval.Duration(),
+		cfg.Ping.Timeout.Duration())
+	monitor.HistorySize = cfg.Ping.History
+
+	err = upsertTargets(desiredTargets, globalResolver, cfg, monitor)
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	go startDNSAutoRefresh(cfg.DNS.Refresh.Duration(), desiredTargets, monitor, cfg)
+	return monitor, nil
+}
+
+func upsertTargets(globalTargets *targets, globalResolver Resolver, cfg *config.Config, monitor *mon.Monitor) error {
+	oldTargets := globalTargets.Targets()
+	newTargets := make([]*target, len(cfg.Targets))
+	var wg sync.WaitGroup
+	var err error
+	for i, t := range cfg.Targets {
+		newTarget := globalTargets.Get(t.Addr)
+		if newTarget == nil {
+			resolver := globalResolver
+			// check if there's a 'resolver' label in the target config
+			// and if its set to 'k8s', then use the k8s resolver
+			if r, ok := t.Labels["resolver"]; ok && r == "k8s" {
+				resolver, err = NewK8sResolver()
+				if err != nil {
+					return fmt.Errorf("failed to create k8s resolver: %w", err)
+				}
+			}
+			newTarget = &target{
+				host:      t.Addr,
+				addresses: make([]net.IPAddr, 0),
+				delay:     time.Duration(10*i) * time.Millisecond,
+				resolver:  resolver,
+			}
+		}
+
+		newTargets[i] = newTarget
+
+		wg.Go(func() {
+			err := newTarget.addOrUpdateMonitor(monitor, targetOpts{
+				disableIPv4: cfg.Options.DisableIPv4,
+				disableIPv6: cfg.Options.DisableIPv6,
+			}, cfg)
+			if err != nil {
+				log.Errorf("failed to setup target: %v", err)
+			}
+		})
+	}
+	wg.Wait()
+	globalTargets.SetTargets(newTargets)
+
+	removed := removedTargets(oldTargets, globalTargets)
+	for _, removedTarget := range removed {
+		log.Infof("remove target: %s", removedTarget.host)
+		removedTarget.removeFromMonitor(monitor)
+	}
+	return nil
+}
+
+func watchConfig(globalTargets *targets, globalResolver Resolver, monitor *mon.Monitor, collector *pingCollector) {
+	watcher, err := inotify.NewWatcher()
+	if err != nil {
+		log.Fatalf("unable to create file watcher: %v", err)
+	}
+
+	err = watcher.Add(*configFile)
+	if err != nil {
+		log.Fatalf("unable to watch file: %v", err)
+	}
+	for {
+		select {
+		case event := <-watcher.Events:
+			log.Debugf("Got file inotify event: %s", event)
+			// If the file is removed, the inotify watcher will lose track of the file. Add it again.
+			if event.Op == inotify.Remove {
+				if err = watcher.Add(*configFile); err != nil {
+					log.Fatalf("failed to renew watch for file: %v", err)
+				}
+			}
+			cfg, err := loadConfig()
+			if err != nil {
+				log.Errorf("unable to load config: %v", err)
+				continue
+			}
+			// We get zero targets if the file was truncated. This happens if an automation tool rewrites
+			// the complete file, instead of alternating only parts of it.
+			if len(cfg.Targets) == 0 {
+				continue
+			}
+			log.Infof("reloading config file %s", *configFile)
+			if err := upsertTargets(globalTargets, globalResolver, cfg, monitor); err != nil {
+				log.Errorf("failed to reload config: %v", err)
+				continue
+			}
+			collector.UpdateConfig(cfg)
+		case err := <-watcher.Errors:
+			log.Errorf("watching file failed: %v", err)
+		}
+	}
+}
+
+func removedTargets(old []*target, new *targets) []*target {
+	var ret []*target
+	for _, oldTarget := range old {
+		if !new.Contains(oldTarget) {
+			ret = append(ret, oldTarget)
+		}
+	}
+	return ret
+}
+
+func startDNSAutoRefresh(interval time.Duration, tar *targets, monitor *mon.Monitor, cfg *config.Config) {
+	if interval <= 0 {
+		return
+	}
+
+	for range time.NewTicker(interval).C {
+		refreshDNS(tar, monitor, cfg)
+	}
+}
+
+func refreshDNS(tar *targets, monitor *mon.Monitor, cfg *config.Config) {
+	log.Infoln("refreshing DNS")
+	for _, t := range tar.Targets() {
+		go func(ta *target) {
+			err := ta.addOrUpdateMonitor(monitor, targetOpts{
+				disableIPv4: cfg.Options.DisableIPv4,
+				disableIPv6: cfg.Options.DisableIPv6,
+			}, cfg)
+			if err != nil {
+				log.Errorf("could not refresh dns: %v", err)
+			}
+		}(t)
+	}
+}
+
+func startServer(collector *pingCollector) {
+	var err error
+	log.Infof("Starting ping exporter (Version: %s)", version)
+	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
+		if !hasValidToken(r, w) {
+			return
+		}
+
+		if _, err := fmt.Fprintf(w, indexHTML, *metricsPath); err != nil {
+			log.Errorf("failed to write response: %v", err)
+		}
+	})
+
+	reg := prometheus.NewRegistry()
+	reg.MustRegister(collector)
+
+	l := log.New()
+	l.Level = log.ErrorLevel
+
+	h := promhttp.HandlerFor(reg, promhttp.HandlerOpts{
+		ErrorLog:      l,
+		ErrorHandling: promhttp.ContinueOnError,
+	})
+	http.HandleFunc(*metricsPath, func(w http.ResponseWriter, r *http.Request) {
+		if !hasValidToken(r, w) {
+			return
+		}
+
+		h.ServeHTTP(w, r)
+	})
+
+	server := http.Server{
+		Addr: *listenAddress,
+	}
+
+	if *serverUseTLS {
+		err = configureTLS(&server)
+		if err != nil {
+			log.Fatalf("could not configure TLS: %v", err)
+		}
+		log.Infof("Listening for %s on %s (HTTPS)", *metricsPath, *listenAddress)
+		err = server.ListenAndServeTLS("", "")
+	} else {
+		log.Infof("Listening for %s on %s (HTTP)", *metricsPath, *listenAddress)
+		err = server.ListenAndServe()
+	}
+
+	if err != nil && err != http.ErrServerClosed {
+		log.Fatal(err)
+	}
+}
+
+func hasValidToken(r *http.Request, w http.ResponseWriter) bool {
+	if *metricsToken == "" {
+		return true
+	}
+
+	token := r.URL.Query().Get("token")
+	if token == "" {
+		token = r.Header.Get("token")
+	}
+
+	if token != *metricsToken {
+		w.WriteHeader(http.StatusForbidden)
+		if _, err := fmt.Fprint(w, "wrong token"); err != nil {
+			log.Errorf("failed to write response: %v", err)
+		}
+		return false
+	}
+
+	return true
+}
+
+func configureTLS(server *http.Server) error {
+	if *serverTLSCertFile == "" || *serverTLSKeyFile == "" {
+		return fmt.Errorf("'web.tls.cert-file' and 'web.tls.key-file' must be defined")
+	}
+
+	server.TLSConfig = &tls.Config{
+		MinVersion: tls.VersionTLS12,
+	}
+
+	var err error
+	server.TLSConfig.Certificates = make([]tls.Certificate, 1)
+	server.TLSConfig.Certificates[0], err = tls.LoadX509KeyPair(*serverTLSCertFile, *serverTLSKeyFile)
+	if err != nil {
+		return fmt.Errorf("loading certificates error: %v", err)
+	}
+
+	if *serverMutualAuthEnabled {
+		err = initMutualAuth(server)
+		if err != nil {
+			return fmt.Errorf("could not initialize mutual auth: %w", err)
+		}
+	} else {
+		server.TLSConfig.ClientAuth = tls.NoClientCert
+	}
+
+	return nil
+}
+
+func initMutualAuth(server *http.Server) error {
+	server.TLSConfig.ClientAuth = tls.RequireAndVerifyClientCert
+
+	if *serverTLSCAFile != "" {
+		var err error
+		var ca []byte
+		if ca, err = os.ReadFile(*serverTLSCAFile); err != nil {
+			return fmt.Errorf("loading CA error: %v", err)
+		} else {
+			server.TLSConfig.ClientCAs = x509.NewCertPool()
+			server.TLSConfig.ClientCAs.AppendCertsFromPEM(ca)
+		}
+	}
+
+	return nil
+}
 
 func loadConfig() (*config.Config, error) {
 	if *configFile == "" {
@@ -167,7 +467,23 @@ func loadConfig() (*config.Config, error) {
 	return cfg, err
 }
 
+func setupGlobalResolver(cfg *config.Config) Resolver {
+	if cfg.DNS.Nameserver == "" {
+		resolver := net.DefaultResolver
+		return resolver
+	}
 
+	if _, _, err := net.SplitHostPort(cfg.DNS.Nameserver); err != nil {
+		// No port present — add default DNS port, handling IPv6 correctly
+		cfg.DNS.Nameserver = net.JoinHostPort(cfg.DNS.Nameserver, "53")
+	}
+	dialer := func(ctx context.Context, _, _ string) (net.Conn, error) {
+		d := net.Dialer{}
+		return d.DialContext(ctx, "udp", cfg.DNS.Nameserver)
+	}
+
+	return &net.Resolver{PreferGo: true, Dial: dialer}
+}
 
 // addFlagToConfig updates cfg with command line flag values, unless the
 // config has non-zero values.
